source clean

+++ src/main/java/org/utils/webcam/SimulatedWebcam.java
@@ -67,7 +67,6 @@ class DummyWebcamDevice implements WebcamDevice
     public DummyWebcamDevice(File f) {
     	setFile(f);
     	init();
-        
     }
 
 	public DummyWebcamDevice(Image ic) {
@@ -162,7 +161,7 @@ class DummyWebcamDevice implements WebcamDevice
         bounceY += yvel*delta;
         if(bounceX > WIDTH - BOUNCE_SIZE)
         {
-            bounceX = WIDTH-BOUNCE_SIZE;
+            bounceX = WIDTH-(double)BOUNCE_SIZE;
             xvel = -xvel;
         }
         if(bounceX < BOUNCE_SIZE)
@@ -172,7 +171,7 @@ class DummyWebcamDevice implements WebcamDevice
         }
         if(bounceY > HEIGHT - BOUNCE_SIZE)
         {
-            bounceY = HEIGHT-BOUNCE_SIZE;
+            bounceY = HEIGHT-(double)BOUNCE_SIZE;
             yvel = -yvel;
         }
         if(bounceY < BOUNCE_SIZE)
